+++ cmd/factory.go
@@ -347,7 +347,7 @@ func newCmdDaemonStatus() *cobra.Command {
 	flags := cmd.Flags()
 	addFlagsGlobal(flags, &options.OptsGlobal)
 	addFlagWatch(flags, &options.Watch)
-	addFlagSections(flags, &options.Sections)
+	addFlagOutputSections(flags, &options.Sections)
 	return cmd
 }
 
@@ -491,7 +491,7 @@ func newCmdMonitor() *cobra.Command {
 	flags := cmd.Flags()
 	addFlagsGlobal(flags, &options.OptsGlobal)
 	addFlagWatch(flags, &options.Watch)
-	addFlagSections(flags, &options.Sections)
+	addFlagOutputSections(flags, &options.Sections)
 	return cmd
 }
 
@@ -1941,13 +1941,17 @@ func newCmdObjectCreate(kind string) *cobra.Command {
 }
 
 func newCmdObjectDelete(kind string) *cobra.Command {
-	// TODO: separate the rid deletion from object deletion. We should have 2 different commands for this
 	var options commands.CmdObjectDelete
 	cmd := &cobra.Command{
 		Use:     "delete",
 		Aliases: []string{"del"},
-		Short:   "delete objects, instances or configuration sections",
-		Long:    "Beware: not setting --local nor --rid deletes all object instances via orchestration, which leaves no local backup of the configuration.",
+		Short:   "delete configuration object or instances (with --local)",
+		Long: "delete configuration object or instances (with --local)\n\n" +
+			"Beware: --local only removes the local instance config." +
+			" The config may be recreated by the daemon from a remote instance copy." +
+			" Without --local the delete is orchestrated so all instance configurations" +
+			" are deleted. The delete command is not responsible for stopping or unprovisioning." +
+			" The deletion happens whatever the object status.",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return options.Run(selectorFlag, kind)
 		},
@@ -1955,7 +1959,6 @@ func newCmdObjectDelete(kind string) *cobra.Command {
 	flags := cmd.Flags()
 	addFlagsGlobal(flags, &options.OptsGlobal)
 	addFlagsLock(flags, &options.OptsLock)
-	addFlagRID(flags, &options.RID)
 	return cmd
 }
 
@@ -2105,7 +2108,7 @@ func newCmdObjectMonitor(kind string) *cobra.Command {
 	flags := cmd.Flags()
 	addFlagsGlobal(flags, &options.OptsGlobal)
 	addFlagWatch(flags, &options.Watch)
-	addFlagSections(flags, &options.Sections)
+	addFlagOutputSections(flags, &options.Sections)
 	return cmd
 }
 
@@ -2668,7 +2671,7 @@ func newCmdObjectUnset(kind string) *cobra.Command {
 	var options commands.CmdObjectUnset
 	cmd := &cobra.Command{
 		Use:   "unset",
-		Short: "unset a configuration key",
+		Short: "unset configuration keywords or sections",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return options.Run(selectorFlag, kind)
 		},
@@ -2677,6 +2680,7 @@ func newCmdObjectUnset(kind string) *cobra.Command {
 	addFlagsGlobal(flags, &options.OptsGlobal)
 	addFlagsLock(flags, &options.OptsLock)
 	addFlagKeywords(flags, &options.Keywords)
+	addFlagSections(flags, &options.Sections)
 	return cmd
 }