+++ CHANGELOG.md
@@ -30,6 +30,10 @@
 
 	users need to use the "unprovision && delete" sequence instead.
 
+* **breaking change:** Remove the --rid flag of the delete command.
+
+  Users can use the "unset --section <name>" command instead.
+
 * **breaking change:** command flags that accept a duration now require a unit.
 
 	change --waitlock=60 to --waitlock=1m
@@ -119,6 +123,8 @@
 * **breaking change:** "om cluster abort" replaces "om node abort" to abort the pending cluster action orchestration.
 
 * **breaking change:** "om ... set|unset" no longer accept --param and --value. Use --kw instead, which was also supported in v2.
+
+* "unset" now accepts "--section <name>" to remove an object configuration section.
 
 * "om monitor" instance availability icons changes: